requirements.txt
argon2-cffi==19.1.0
asn1crypto==0.24.0
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
cryptography==2.7
Django==2.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-debug-toolbar==2.0
fabric2==2.5.0
Fabric3==1.14.post1
Faker==1.0.7
invoke==1.3.0
paramiko==2.6.0
Pillow==6.0.0
pycparser==2.19
PyNaCl==1.3.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
virtualenv==16.7.2
wincertstore==0.2
